Refactored AggregationDataStoreHelper class
