+++ elide-datastore/elide-datastore-aggregation/src/main/java/com/yahoo/elide/datastores/aggregation/AggregationDataStoreTransaction.java
@@ -65,8 +65,8 @@ public class AggregationDataStoreTransaction implements DataStoreTransaction {
     Query buildQuery(EntityProjection entityProjection, RequestScope scope) {
         Schema schema = queryEngine.getSchema(entityProjection.getType());
 
-        AggregationDataStoreHelper agHelper = new AggregationDataStoreHelper(schema,
+        EntityProjectionTranslator translator = new EntityProjectionTranslator(schema,
                 entityProjection, scope.getDictionary());
-        return agHelper.getQuery();
+        return translator.getQuery();
     }
 }
