+++ elide-datastore/elide-datastore-aggregation/src/main/java/com/yahoo/elide/datastores/aggregation/EntityProjectionTranslator.java
@@ -6,14 +6,8 @@
 package com.yahoo.elide.datastores.aggregation;
 
 import com.yahoo.elide.core.EntityDictionary;
-import com.yahoo.elide.core.Path;
 import com.yahoo.elide.core.exceptions.InvalidOperationException;
-import com.yahoo.elide.core.filter.FilterPredicate;
-import com.yahoo.elide.core.filter.expression.AndFilterExpression;
 import com.yahoo.elide.core.filter.expression.FilterExpression;
-import com.yahoo.elide.core.filter.expression.NotFilterExpression;
-import com.yahoo.elide.core.filter.expression.OrFilterExpression;
-import com.yahoo.elide.core.sort.Sorting;
 import com.yahoo.elide.datastores.aggregation.annotation.TimeGrainDefinition;
 import com.yahoo.elide.datastores.aggregation.filter.visitor.FilterConstraints;
 import com.yahoo.elide.datastores.aggregation.filter.visitor.SplitFilterExpressionVisitor;
@@ -33,19 +27,14 @@ import com.yahoo.elide.request.Relationship;
 
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
-import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-/**
- * Helper for Aggregation Data Store which does the work associated with extracting {@link Query}.
- */
-public class AggregationDataStoreHelper {
+public class EntityProjectionTranslator {
 
-    //TODO refactor this class in the next PR
     //TODO Add support for user selected metrics.
     private static final int AGGREGATION_METHOD_INDEX = 0;
 
@@ -58,7 +47,7 @@ public class AggregationDataStoreHelper {
     private FilterExpression havingFilter;
     private EntityDictionary dictionary;
 
-    public AggregationDataStoreHelper(Schema schema, EntityProjection entityProjection, EntityDictionary dictionary) {
+    public EntityProjectionTranslator(Schema schema, EntityProjection entityProjection, EntityDictionary dictionary) {
         this.dictionary = dictionary;
         this.schema = schema;
         this.entityProjection = entityProjection;
@@ -68,7 +57,6 @@ public class AggregationDataStoreHelper {
         resolveMetricList(metrics);
         metricMap = new LinkedHashMap<>();
         resolveMetricMap(metrics);
-        validateSorting();
         splitFilters();
     }
 
@@ -77,7 +65,7 @@ public class AggregationDataStoreHelper {
      * @return {@link Query} query object with all the parameters provided by user.
      */
     public Query getQuery() {
-        return Query.builder()
+        Query query =  Query.builder()
                 .schema(schema)
                 .metrics(metricMap)
                 .groupDimensions(dimensionProjections)
@@ -87,6 +75,9 @@ public class AggregationDataStoreHelper {
                 .sorting(entityProjection.getSorting())
                 .pagination(entityProjection.getPagination())
                 .build();
+        QueryValidator validator = new QueryValidator(getAllFields(), query, dictionary, entityProjection.getType());
+        validator.validate();
+        return query;
     }
 
     /**
@@ -103,10 +94,6 @@ public class AggregationDataStoreHelper {
         FilterConstraints constraints = filterExpression.accept(visitor);
         whereFilter = constraints.getWhereExpression();
         havingFilter = constraints.getHavingExpression();
-
-        if (havingFilter != null) {
-            validateHavingClause(havingFilter);
-        }
     }
 
     /**
@@ -218,95 +205,9 @@ public class AggregationDataStoreHelper {
     }
 
     /**
-     * Validate the having clause before execution. Having clause is not as flexible as where clause,
-     * the fields in having clause must be either or these two:
-     * 1. A grouped by dimension in this query
-     * 2. An aggregated metric in this query
-     *
-     * All grouped by dimensions are defined in the entity bean, so the last entity class of a filter path
-     * must match entity class of the query.
-     *
-     * @param havingClause having clause generated from this query
+     * Helper method to get all fields from the {@link EntityProjection}.
+     * @return allFields set of all field names
      */
-    private void validateHavingClause(FilterExpression havingClause) {
-        if (havingClause instanceof FilterPredicate) {
-            Path path = ((FilterPredicate) havingClause).getPath();
-            Path.PathElement last = path.lastElement().get();
-            Class<?> cls = last.getType();
-            String field = last.getFieldName();
-
-            if (cls != schema.getEntityClass()) {
-                throw new InvalidOperationException(
-                        String.format(
-                                "Classes don't match when try filtering on %s in having clause of %s.",
-                                cls.getSimpleName(),
-                                schema.getEntityClass().getSimpleName()));
-            }
-
-            if (schema.isMetricField(field)) {
-                Metric metric = schema.getMetric(field);
-                if (!metricMap.containsKey(metric)) {
-                    throw new InvalidOperationException(
-                            String.format(
-                                    "Metric field %s must be aggregated before filtering in having clause.", field));
-                }
-            } else {
-                if (dimensionProjections.stream().noneMatch(dim -> dim.getName().equals(field))) {
-                    throw new InvalidOperationException(
-                            String.format(
-                                    "Dimension field %s must be grouped before filtering in having clause.", field));
-                }
-            }
-        } else if (havingClause instanceof AndFilterExpression) {
-            validateHavingClause(((AndFilterExpression) havingClause).getLeft());
-            validateHavingClause(((AndFilterExpression) havingClause).getRight());
-        } else if (havingClause instanceof OrFilterExpression) {
-            validateHavingClause(((OrFilterExpression) havingClause).getLeft());
-            validateHavingClause(((OrFilterExpression) havingClause).getRight());
-        } else if (havingClause instanceof NotFilterExpression) {
-            validateHavingClause(((NotFilterExpression) havingClause).getNegated());
-        }
-    }
-
-    /**
-     * Method to verify that all the sorting options provided
-     * by the user are valid and supported.
-     */
-    public void validateSorting() {
-        Sorting sorting = entityProjection.getSorting();
-        if (sorting == null) {
-            return;
-        }
-        Set<String> allFields = getAllFields();
-        Map<Path, Sorting.SortOrder> sortClauses = sorting.getValidSortingRules(entityProjection.getType(), dictionary);
-        sortClauses.keySet().forEach((path) -> validateSortingPath(path, allFields));
-    }
-
-    /**
-     * Verifies that the current path can be sorted on
-     * @param path The path that we are validating
-     * @param allFields Set of all field names included in initial query
-     */
-    private void validateSortingPath(Path path, Set<String> allFields) {
-        List<Path.PathElement> pathElemenets = path.getPathElements();
-
-        //TODO add support for double nested sorting
-        if (pathElemenets.size() > 2) {
-            throw new UnsupportedOperationException(
-                    "Currently sorting on double nested fields is not supported");
-        }
-        Path.PathElement currentElement = pathElemenets.get(0);
-        String currentField = currentElement.getFieldName();
-        Class<?> currentClass = currentElement.getType();
-        if (!allFields.stream().anyMatch(field -> field.equals(currentField))) {
-            throw new InvalidOperationException("Can't sort on " + currentField + " as it is not present in query");
-        }
-        if (dictionary.getIdFieldName(currentClass).equals(currentField)
-                || currentField.equals(EntityDictionary.REGULAR_ID_NAME)) {
-            throw new InvalidOperationException("Sorting on id field is not permitted");
-        }
-    }
-
     private Set<String> getAllFields() {
         Set<String> allFields = getAttributes();
         allFields.addAll(getRelationships());
